+++ src/libponyc/ast/token.c
@@ -364,7 +364,7 @@ static void token_signature_serialise_trace(pony_ctx_t* ctx, void* object)
 }
 
 static void token_signature_serialise(pony_ctx_t* ctx, void* object, void* buf,
-  size_t offset, int mutability)
+  uint64_t offset, int mutability)
 {
   (void)mutability;
 
@@ -437,7 +437,7 @@ static void token_docstring_signature_serialise_trace(pony_ctx_t* ctx,
 }
 
 static void token_docstring_signature_serialise(pony_ctx_t* ctx, void* object,
-  void* buf, size_t offset, int mutability)
+  void* buf, uint64_t offset, int mutability)
 {
   (void)ctx;
   (void)object;
@@ -490,7 +490,7 @@ static void token_serialise_trace(pony_ctx_t* ctx, void* object)
 }
 
 static void token_serialise(pony_ctx_t* ctx, void* object, void* buf,
-  size_t offset, int mutability)
+  uint64_t offset, int mutability)
 {
   (void)mutability;
