+++ src/libponyc/pkg/program.c
@@ -358,7 +358,7 @@ static void program_serialise_trace(pony_ctx_t* ctx, void* object)
 }
 
 static void program_serialise(pony_ctx_t* ctx, void* object, void* buf,
-  size_t offset, int mutability)
+  uint64_t offset, int mutability)
 {
   (void)mutability;
